# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Pachi API
  version: 0.1.0
  description: Identity foundation, private individual-provider marketplace operations, and bounded organization onboarding/ordinary membership governance. Organization activation remains separate from BUSINESS verification and publication eligibility.
paths:
...
  /v1/account/listings/{listingId}/inquiry:
    parameters: [{ name: listingId, in: path, required: true, schema: { type: string, format: uuid } }]
    post:
      security: [{ bearerAuth: [] }]
      description: Creates or reuses one OPEN Interaction and its Conversation for the authenticated phone-verified seeker. An existing RESTRICTED context returns a safe 403; CLOSED history may be followed by a new context when all current listing eligibility rules pass. Provider context is derived from the listing and the request body is empty.
      parameters: [{ name: Idempotency-Key, in: header, required: true, schema: { type: string, format: uuid } }]
      responses: { '201': { description: New or reused interaction context, content: { application/json: { schema: { $ref: '#/components/schemas/InteractionResponse' } } } }, '401': { description: Authentication required }, '403': { description: Phone verification required or contact capability restricted; no internal restriction details are returned }, '404': { description: Listing is not publicly available }, '409': { description: Invalid inquiry or idempotency key conflict } }
...
components:
  securitySchemes:
    staffBearerAuth: { type: http, scheme: bearer, bearerFormat: JWT, description: Separate staff issuer/client access token with pachi/staff scope }
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
...
    InteractionResponse:
      type: object
      required: [interaction_id, conversation_id, listing_id, state, opened_at, created]
      properties:
        interaction_id: { type: string, format: uuid }
        conversation_id: { type: string, format: uuid }
        listing_id: { type: string, format: uuid }
        state: { type: string, enum: [OPEN, CLOSED, RESTRICTED] }
        opened_at: { type: string, format: date-time }
        created: { type: boolean }
...